Move TokenKinds.def.gyb to AST
